Add files via upload
To ensure my code is functioning properly is by using automated test to increase efficiency and accuracy. CI/CD can also be used when there are updates. I would use the help of a peer to check the code has no errors and check it myself after all is done to give it a final check.

To interpret the needs of the user, I would get a clear understanding of the requirements and seek constant feedback to tailor the program to the user's needs.

In software design, I start with the planning of the program.  I would then create this code in parts and use an iteration approach to allow changes to be more efficient.